Sets Zohilan as CLIP's official language (#5536)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Sets Zohilan as CLIP's official language, because I forgot to when
adding it to the game.

## Why It's Good For The Game

CLIP's capital and industrial core are former Zohil colonies and should
have this as their official language.

## Changelog

:cl:
fix: CLIP now uses Zohilan as its official language
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->