Cobra changes and small tweaks for others (#5532)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Changes the indie Cobra burst firemode to be usefull (it's been rarely
used due to it's condition of past 5 months). Adjusts the spread of SMGs
and PDWs (both Cobras, Resolution and Mongrel) to match their
counterparts.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes indie Cobra playable again.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: changes civ Cobra to rapid fire bursts of 4. Decreases spread
on Cobras to 2, (3-1 from silencer on mil cobra) and increases on
Mongrel and Resolution by 1
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->