Adds a few medkit sprites, and adds the SRM medkit to cargo. (#5507)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
This PR adds an icon and inhand for the SRM Medkit, slightly changes the
Advanced Medkit icon and inhand, adds an icon for the Ancient Medkit,
changes the pill bottle's inhand, and changes several descriptions to
make them more descriptive.

Also moves the Roumain medkit out of the Elder's map file(why the hell
was it stored there) and with the rest of the medkits, as well as adds
it to cargo for 400 credits, with a 10% discount for the SRM.

Also also: Removes the mushroom bowl out of the SRM medkit since it was
normal-sized and shouldn't be able to fit in there. Fixes ALL medkits to
be able to have stuff put back into them after you take it out.

<details><summary>Images:</summary>
<p>
<img width="423" height="303" alt="Screenshot (4607)"
src="https://github.com/user-attachments/assets/aede6804-504e-4788-8e21-c0c30bdef218"
/>
<img width="408" height="285" alt="Screenshot (4608)"
src="https://github.com/user-attachments/assets/7fdc12bf-8809-4922-b651-b13b7ac425ce"
/>
</p>
</details>
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
The Roumain kit used the same sprite as the Advanced kit, and I wanted
to make the Advanced kit stand out a bit more. Likewise, the ancient kit
used the same as the standard first aid kit, and I think making a more
"decrepit" looking FA kit would look good in certain ruins. Pill bottle
also looked weird and too dark.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: Pill bottle and several medkits' sprites and descriptions updated.
add: Roumain medkit is now purchasable at the outpost.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Aquidu <98631832+Aquidu@users.noreply.github.com>