Adds a blood filter to cargo (#5635)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a purchasable blood filter to cargo for 500 credits. There are no
discounts.

## Why It's Good For The Game
The blood filter is currently inaccessible to players even though the
surgery is in the game. The blood filter is an item that purges anything
not on its whitelist from a body. It lets you purge toxins/overdoses
from people without purgatives in a manner, even in corpses. There's not
really anything like it. Price is 500 credits because the defib is 750
credits.
## Changelog

:cl: Goat
add: Blood filter in cargo for 500 credits. 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: Erika Fox <94164348+Erikafox@users.noreply.github.com>