fix neoxathin deplete + add overdose effect (#5640)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

adds missing `..()` to the end of `on_mob_life()` of nexoathin and gives
it an overdose effect where you have a 30% chance to vomit

## Why It's Good For The Game
fixes nexoathin not depleteting and being stuck in the body forever
permanently healing toxin

gives nexoathin a proper OD effect where you vomit because this will
(most likely) be consumed orally as it's a flora medicine and also used
in soups so your body will try to vomit it out (because it felt bad for
it to just be a mood debuff and not do anything else when it had a
defined overdose amount).
## Changelog

:cl: Goat
fix: nexoathin depletes now
balance: nexoathin makes you vomit now when you overdose on it
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Goat <126099705+Goat-Real@users.noreply.github.com>
Co-authored-by: Erika Fox <94164348+Erikafox@users.noreply.github.com>