Fixes being unable to place certain tools inside crates or lockers (#5690)
Minor oversight. If the crate/locker is open it now requires
right-clicking for the tool's special interaction, making it possible to
place the tool inside using left click.

:cl:
fix: Fixed being unable to place certain tools inside crates or lockers.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->