makes the bluespace drive fluff machine dense and fixes a typo in its desc (#5706)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
fixes #5689 
makes it dense and correctly-spelled

## Why It's Good For The Game

its a GIANT ASS MACHINE THE SIZE OF A THRUSTER. all the interior design
featuring it assumes its dense and the lack of such is breaking my
immersion (also spellcheck good)

## Changelog

:cl:
fix: Bluespace drive fluff machines are now dense
spellcheck: Corrected a typo in bluespace drive machine description.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Addust <80979251+Addust@users.noreply.github.com>
Co-authored-by: Erika Fox <94164348+Erikafox@users.noreply.github.com>