Allows water to be used to reduce disgust (#5734)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Much in the same way as sol dry, this PR makes it so that a glass of
water can be used to quell nausea

## Why It's Good For The Game

I'm not totally sure if this should be as strong as sol dry, but I felt
like it made for fun flavor for a cup of water to help with queasiness
and all

## Changelog

:cl:
add: Added the ability for water to be used to reduce disgust levels
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->